requirements.txt
einops==0.4.1
wandb==0.12.17
torch==1.11.0
Pillow==7.1.2
pandas==1.3.5
opencv-python==4.1.2.30
utils
numpy
matplotlib